Reorganize autopilot package
